fix(wan2.2): preserve BF16 dtype in RMSNorm

+++ turbodiffusion/rcm/networks/wan2pt2.py
@@ -205,7 +205,7 @@ class WanRMSNorm(nn.Module):
         Args:
             x(Tensor): Shape [B, L, C]
         """
-        return self._norm(x.float()).type_as(x) * self.weight
+        return (self._norm(x.float()).type_as(x) * self.weight).type_as(x)
 
     def _norm(self, x):
         return x * torch.rsqrt(x.pow(2).mean(dim=-1, keepdim=True) + self.eps)
